refactor: WaitingExistsForSlotException을 service/exception으로 이동
